Separated classification and regression strategies
