+++ active_learning/active_learning.py
@@ -1,9 +1,10 @@
 import logging
-import numpy as np
 from collections import Iterable
+
+import numpy as np
 from tqdm import tqdm
 
-from active_learning.query_strats import uncertainty_sampling
+from active_learning.query_strats.classification import uncertainty_sampling
 from active_learning.selectors import identity_selector
