+++ active_learning/query_strats/classification/__init__.py
@@ -1,0 +1,1 @@
+"""Query strategies specific to classification problems"""
