+++ active_learning/query_strats/classification/tests/test_greedy.py
@@ -1,8 +1,10 @@
-from ..greedy import GreedySearch
-from ...tests.test_problem import make_grid_problem
-from sklearn.neighbors import KNeighborsClassifier
 from unittest import TestCase
+
 import numpy as np
+from sklearn.neighbors import KNeighborsClassifier
+
+from active_learning.query_strats.classification.greedy import GreedySearch
+from active_learning.tests.test_problem import make_grid_problem
 
 
 class TestGreedy(TestCase):
