+++ active_learning/query_strats/classification/uncertainty_sampling.py
@@ -1,5 +1,5 @@
-from ..problem import ActiveLearningProblem
-from . import IndividualScoreQueryStrategy
+from active_learning.problem import ActiveLearningProblem
+from active_learning.query_strats import IndividualScoreQueryStrategy
 from typing import List
 import numpy as np
