+++ active_learning/query_strats/regression/__init__.py
@@ -1,0 +1,1 @@
+"""Query strategies specific to regression problems"""
